+++ controllers/apply/awards-for-all/fields.js
@@ -33,6 +33,7 @@ const fieldProjectDateRange = require('./fields/project-date-range');
 const fieldCompanyNumber = require('./fields/company-number');
 const fieldCharityNumber = require('./fields/charity-number');
 const fieldEducationNumber = require('./fields/education-number');
+const fieldPhone = require('./fields/phone');
 
 module.exports = function fieldsFor({ locale, data = {} }) {
     const localise = get(locale);
@@ -98,35 +99,6 @@ module.exports = function fieldsFor({ locale, data = {} }) {
                     })
                 }
             ].concat(additionalMessages)
-        };
-
-        return { ...defaultProps, ...props };
-    }
-
-    function phoneField(props) {
-        const defaultProps = {
-            type: 'tel',
-            attributes: { size: 30, autocomplete: 'tel' },
-            isRequired: true,
-            schema: Joi.string()
-                .phoneNumber()
-                .required(),
-            messages: [
-                {
-                    type: 'base',
-                    message: localise({
-                        en: 'Enter a UK telephone number',
-                        cy: 'Rhowch rif ffôn Prydeinig'
-                    })
-                },
-                {
-                    type: 'string.phonenumber',
-                    message: localise({
-                        en: 'Enter a real UK telephone number',
-                        cy: 'Rhowch rif ffôn Prydeinig go iawn'
-                    })
-                }
-            ]
         };
 
         return { ...defaultProps, ...props };
@@ -1647,9 +1619,8 @@ module.exports = function fieldsFor({ locale, data = {} }) {
                 }
             ]
         ),
-        mainContactPhone: phoneField({
-            name: 'mainContactPhone',
-            label: localise({ en: 'Telephone number', cy: 'Rhif ffôn' })
+        mainContactPhone: fieldPhone(locale, {
+            name: 'mainContactPhone'
         }),
         mainContactLanguagePreference: fieldContactLanguagePreference(locale, {
             name: 'mainContactLanguagePreference'
@@ -1745,9 +1716,8 @@ module.exports = function fieldsFor({ locale, data = {} }) {
                     'Byddwn yn defnyddio hwn pan fyddwn yn cysylltu ynglŷn â’r prosiect'
             })
         }),
-        seniorContactPhone: phoneField({
-            name: 'seniorContactPhone',
-            label: localise({ en: 'Telephone number', cy: 'Rhif ffôn' })
+        seniorContactPhone: fieldPhone(locale, {
+            name: 'seniorContactPhone'
         }),
         seniorContactLanguagePreference: fieldContactLanguagePreference(
             locale,
